Remove using byteValue() not needed

+++ db/cassandra-0.8/src/com/yahoo/ycsb/db/CassandraClient8.java
@@ -507,7 +507,7 @@ public class CassandraClient8 extends DB
             for (String field : values.keySet()) {
                 ByteBuffer value;
                 if (timeseries)
-                    value = ByteBufferUtil.bytes(new Long(System.currentTimeMillis() / 1000L).byteValue());
+                    value = ByteBufferUtil.bytes(new Long(System.currentTimeMillis() / 1000L));
                 else
                     value = tob(field);
